+++ packages/codegen/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/codegen [1.0.0-alpha.49](https://github.com/anolilab/lunora/compare/%40lunora%2Fcodegen%401.0.0-alpha.48...%40lunora%2Fcodegen%401.0.0-alpha.49) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/do:** upgraded to 1.0.0-alpha.37
+
 ## @lunora/codegen [1.0.0-alpha.48](https://github.com/anolilab/lunora/compare/%40lunora%2Fcodegen%401.0.0-alpha.47...%40lunora%2Fcodegen%401.0.0-alpha.48) (2026-07-21)