+++ packages/config/CHANGELOG.md
@@ -1,3 +1,17 @@
+## @lunora/config [1.0.0-alpha.75](https://github.com/anolilab/lunora/compare/%40lunora%2Fconfig%401.0.0-alpha.74...%40lunora%2Fconfig%401.0.0-alpha.75) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/codegen:** upgraded to 1.0.0-alpha.49
+
+## @lunora/config [1.0.0-alpha.74](https://github.com/anolilab/lunora/compare/%40lunora%2Fconfig%401.0.0-alpha.73...%40lunora%2Fconfig%401.0.0-alpha.74) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/studio:** upgraded to 1.0.0-alpha.57
+
 ## @lunora/config [1.0.0-alpha.73](https://github.com/anolilab/lunora/compare/%40lunora%2Fconfig%401.0.0-alpha.72...%40lunora%2Fconfig%401.0.0-alpha.73) (2026-07-21)