+++ packages/lunora/CHANGELOG.md
@@ -1,3 +1,21 @@
+## lunorash [1.0.0-alpha.95](https://github.com/anolilab/lunora/compare/lunorash%401.0.0-alpha.94...lunorash%401.0.0-alpha.95) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/cli:** upgraded to 1.0.0-alpha.94
+* **@lunora/do:** upgraded to 1.0.0-alpha.37
+* **@lunora/runtime:** upgraded to 1.0.0-alpha.33
+
+## lunorash [1.0.0-alpha.94](https://github.com/anolilab/lunora/compare/lunorash%401.0.0-alpha.93...lunorash%401.0.0-alpha.94) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/cli:** upgraded to 1.0.0-alpha.93
+* **@lunora/client:** upgraded to 1.0.0-alpha.26
+* **@lunora/runtime:** upgraded to 1.0.0-alpha.32
+
 ## lunorash [1.0.0-alpha.93](https://github.com/anolilab/lunora/compare/lunorash%401.0.0-alpha.92...lunorash%401.0.0-alpha.93) (2026-07-21)