+++ packages/react/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/react [1.0.0-alpha.30](https://github.com/anolilab/lunora/compare/%40lunora%2Freact%401.0.0-alpha.29...%40lunora%2Freact%401.0.0-alpha.30) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.26
+
 ## @lunora/react [1.0.0-alpha.29](https://github.com/anolilab/lunora/compare/%40lunora%2Freact%401.0.0-alpha.28...%40lunora%2Freact%401.0.0-alpha.29) (2026-07-20)