+++ packages/studio/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @lunora/studio [1.0.0-alpha.57](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.56...%40lunora%2Fstudio%401.0.0-alpha.57) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.26
+* **@lunora/react:** upgraded to 1.0.0-alpha.30
+
 ## @lunora/studio [1.0.0-alpha.56](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.55...%40lunora%2Fstudio%401.0.0-alpha.56) (2026-07-21)
 
 ## @lunora/studio [1.0.0-alpha.55](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.54...%40lunora%2Fstudio%401.0.0-alpha.55) (2026-07-21)