+++ packages/vite/CHANGELOG.md
@@ -1,3 +1,19 @@
+## @lunora/vite [1.0.0-alpha.75](https://github.com/anolilab/lunora/compare/%40lunora%2Fvite%401.0.0-alpha.74...%40lunora%2Fvite%401.0.0-alpha.75) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/codegen:** upgraded to 1.0.0-alpha.49
+* **@lunora/config:** upgraded to 1.0.0-alpha.75
+
+## @lunora/vite [1.0.0-alpha.74](https://github.com/anolilab/lunora/compare/%40lunora%2Fvite%401.0.0-alpha.73...%40lunora%2Fvite%401.0.0-alpha.74) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/config:** upgraded to 1.0.0-alpha.74
+* **@lunora/studio:** upgraded to 1.0.0-alpha.57
+
 ## @lunora/vite [1.0.0-alpha.73](https://github.com/anolilab/lunora/compare/%40lunora%2Fvite%401.0.0-alpha.72...%40lunora%2Fvite%401.0.0-alpha.73) (2026-07-21)